Create Table.js component
In this component are received data from parent and new state is created from them.
Here are rendered table rows by renderTableRows function, which map the state and return
TableRow component for each of objects.
In table headers are added onClick events which run sortByRows function with parameter depend on clicked element.